requirements.txt
django
Markdown
django-filter
djangorestframework
psycopg2-binary
django-ckeditor
Pillow